requirements.txt
absl-py==2.1.0
accelerate==1.2.1
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
filelock==3.16.1
flatbuffers==1.12
fsspec==2024.12.0
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
protobuf==3.19.6
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.2
six==1.15.0
sympy==1.13.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.9.0
tensorflow-metal==0.5.0
termcolor==2.5.0
tokenizers==0.21.0
torch==2.2.2
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
